requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
fonttools==4.46.0
fsspec==2023.4.0
fvcore==0.1.5.post20221221
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
nas-bench-graph==1.4.0
networkx==3.0
numpy==1.24.1
packaging==23.2
Pillow==9.3.0
portalocker==2.8.2
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
pywin32==306
PyYAML==6.0.1
requests==2.28.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.2.0
torch==2.1.1+cu121
torch_geometric==2.4.0
torchaudio==2.1.1+cu121
torchvision==0.16.1+cu121
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.13
yacs==0.1.8
